VP_Brain/requirements.txt
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.6
requests>=2.31
websockets>=12.0
python-multipart>=0.0.9
python-dotenv>=1.0
google-generativeai>=0.7
httpx>=0.27
